Using flash messages to display feedback to the user... plus a lot of other usability changes.

+++ clean.py
@@ -8,5 +8,5 @@ connect('', host=env['registry'].settings.get('mongodb', 'packassembler'))
 too_old = datetime.now() - timedelta(days=45)
 
 for user in User.objects:
-    if not Mod.objects(owner=user) and user.group == 'user' and (not user.last_login or user.last_login < too_old):
+    if not Mod.objects(owner=user) and user.group == 'user' and (not user.last_login or user.last_login < too_old) and user.username != 'Orphan':
         user.delete()
